requirements.txt
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
greenlet==3.2.3
idna==3.10
isodate==0.7.2
numpy==2.2.6
pandas==2.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
